requirements.txt
feedparser>=6.0.11
PyYAML>=6.0.2
requests>=2.32.0
